// src/scripts/constants/interfaces/proficiencyprops.ts
import { ProficiencySkillProps } from "./proficiencyskillprops";

export interface ProficiencyProps {
  cooking: ProficiencySkillProps;
  crafting: ProficiencySkillProps;
  fishing: ProficiencySkillProps;
  farming: ProficiencySkillProps;
  herbGathering: ProficiencySkillProps;
  livestockTending: ProficiencySkillProps;
  metalcrafting: ProficiencySkillProps;
  mining: ProficiencySkillProps;
  researching: ProficiencySkillProps;
  woodchopping: ProficiencySkillProps;
}